1 - dodata klasa porudzbina sa ogranicenjima 2 - svi nasleduju klasu indentifikator da bi imali atribut id koji ce biti koriscen prilikom cuvanja objekata u fajlove TODO: redefinisati equals metodu restorana da bi validacija u porudzbini radila

+++ src/main/java/ftn/vu/model/Dostavljac.java
@@ -10,7 +10,7 @@ public class Dostavljac extends Korisnik {
 	
 	private String registarskaOznakaVozila;
 	
-	public int getJmbg() {
+	public long getJmbg() {
 		return jmbg;
 	}
